STYLE:Fixed the doc about the size of VTK_BIT type. BUG:Made the static version and the member version of GetDataTypeSize() have the same return type

+++ VTK/Common/vtkAbstractArray.cxx
@@ -97,12 +97,12 @@ void vtkAbstractArray::GetTuples(vtkIdType p1, vtkIdType p2,
 
 //----------------------------------------------------------------------------
 template <class T>
-unsigned long vtkAbstractArrayGetDataTypeSize(T*)
+int vtkAbstractArrayGetDataTypeSize(T*)
 {
   return sizeof(T);
 }
 
-unsigned long vtkAbstractArray::GetDataTypeSize(int type)
+int vtkAbstractArray::GetDataTypeSize(int type)
 {
   switch (type)
     {
